requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
dj-database-url>=2.1.0
python-fasthtml>=0.1.0
